Rename and reogarnise test files
